Factor our DistinctList into a separate module
